requirements.txt
alembic==1.11.1
certifi==2023.5.7
charset-normalizer==3.1.0
databricks-connect==13.1.0
databricks-sdk==0.1.9
databricks-sql-connector==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
iniconfig==2.0.0
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pluggy==1.0.0
protobuf==4.23.2
py4j==0.10.9.7
pyarrow==12.0.0
pytest==7.3.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.48
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
